requirements.txt
pandas==2.3.3
openpyxl==3.1.5
requests==2.32.5
python-dotenv==1.0.0
numpy>=1.24.0
fuzzywuzzy==0.18.0
python-levenshtein==0.27.3
# chromadb>=0.4.22
google-generativeai>=0.8.3